ran tests against endpoints
