+++ model/courses.js
@@ -1,15 +1,6 @@
-const { ObjectId, GridFSBucket } = require('mongodb')
+const { ObjectId } = require('mongodb')
 const { getDbReference } = require('../lib/mongo')
-const auth = require('../lib/auth');
 
-const CoursesSchema = {
-    subjectCode: {required: true},
-    number: {required: true},
-    title: {required: true},
-    term: {required: true},
-    instructorId: {required: true},
-    students: {required: true}
-}
 
 
 /*
@@ -20,13 +11,12 @@ async function delete_course(courseId) {
 
     const db = getDbReference();
     const courseCollection = db.collection("Courses");
-    const courseResult = await courseCollection.deleteOne({ _id: courseId });
+    const courseResult = await courseCollection.deleteOne({ _id: new ObjectId(String(courseId)) });
     if(courseResult.acknowledged && courseResult.deletedCount === 1){
         const assignmentCollection = db.collection("Assignments");
-        const assignmentResult= await assignmentCollection.deleteMany({ courseId: courseId });
-        if(assignmentResult.acknowledged && assignmentResult.deletedCount === 1){
-            return 0;
-        }
+        assignmentCollection.deleteMany({ courseId: new ObjectId(String(courseId)) });
+
+        return 0;
     }
     return -1;
 }
@@ -39,7 +29,7 @@ exports.delete_course = delete_course
 async function get_courses_by_id(courseId){
     const db = getDbReference();
     const collection = db.collection("Courses");
-    const result = await collection.findOne({ _id: courseId });
+    const result = await collection.findOne({ _id: new ObjectId(String(courseId)) });
     return result;
 }
 exports.get_courses_by_id = get_courses_by_id
@@ -50,19 +40,21 @@ exports.get_courses_by_id = get_courses_by_id
 async function update_students(req, courseId){
     const db = getDbReference();
     const collection = db.collection("Courses");
-
+    
     const toAdd = req.body.add;
-    if(toAdd.length > 0){
-        db.collection.updateOne(
-            { _id: courseId },
+
+    if(toAdd & toAdd.length > 0){
+        await collection.updateOne(
+            { _id: new ObjectId(String(courseId)) },
             { $addToSet: { students: { $each: toAdd } } }
         );
     }
+
     const toRemove = req.body.remove;
-    if(toRemove.length > 0){
-        db.collection.updateOne(
-            { _id: courseId },
-            { $pullAll: { students: { $each: toRemove } } }
+    if(toRemove && toRemove.length > 0){
+        await collection.updateOne(
+            { _id: new ObjectId(String(courseId)) },
+            { $pull: { students: { $in: toRemove } } }
         );
     }
